requirements.txt.txt
numpy>=1.24
pillow>=10.0
tqdm>=4.66
matplotlib>=3.8
scikit-learn>=1.4
pandas>=2.2
transformers>=4.44
torch>=2.2
torchvision>=0.17
huggingface-hub>=0.24
requests>=2.32
